feat: initial commit with templated code in place
CADC-14318
