+++ config/config.yml
@@ -8,7 +8,7 @@ proxy_file_name: /usr/src/app/cadcproxy.pem
 # operational value is ivo://cadc.nrc.ca/ams
 # - modifies entries on sc2.canfar.net
 resource_id: ivo://cadc.nrc.ca/sc2repo
-# operational value is ivo://cadc.nrc.ca/ams/blank
+# operational value is ivo://cadc.nrc.ca/ams/lsst
 # - queries entries on sc2.canfar.net
 tap_id: ivo://cadc.nrc.ca/sc2tap
 storage_inventory_resource_id: ivo://cadc.nrc.ca/cadc/minoc
@@ -50,7 +50,7 @@ retry_file_name: retries.txt
 # this file is created in the logs directory, default is 'progress.txt'
 progress_file_name: progress.txt
 #
-collection: BLANK
+collection: LSST
 scheme: cadc
 preview_scheme: cadc
 #
