+++ lsst2caom2/file2caom2_augmentation.py
@@ -2,7 +2,7 @@
 # ******************  CANADIAN ASTRONOMY DATA CENTRE  *******************
 # *************  CENTRE CANADIEN DE DONNÉES ASTRONOMIQUES  **************
 #
-#  (c) 2021.                            (c) 2021.
+#  (c) 2025.                            (c) 2025.
 #  Government of Canada                 Gouvernement du Canada
 #  National Research Council            Conseil national de recherches
 #  Ottawa, Canada, K1A 0R6              Ottawa, Canada, K1A 0R6
@@ -68,21 +68,21 @@
 
 
 from caom2pipe import caom_composable as cc
-from blank2caom2 import main_app
+from lsst2caom2 import main_app
 
 
-__all__ = ['BlankFits2caom2Visitor']
+__all__ = ['LSSTFits2caom2Visitor']
 
 
-class BlankFits2caom2Visitor(cc.Fits2caom2Visitor):
+class LSSTFits2caom2Visitor(cc.Fits2caom2VisitorRunnerMeta):
     def __init__(self, observation, **kwargs):
         super().__init__(observation, **kwargs)
 
-    def _get_mapping(self, headers, _):
-        return main_app.BlankMapping(
-            self._storage_name, headers, self._clients, self._observable, self._observation, self._config
-        )
+    def _get_mappings(self, dest_uri):
+        return [main_app.LSSTMapping(
+            self._storage_name, self._clients, self._reporter, self._observation, self._config
+        )]
 
 
 def visit(observation, **kwargs):
-    return BlankFits2caom2Visitor(observation, **kwargs).visit()
+    return LSSTFits2caom2Visitor(observation, **kwargs).visit()
