+++ scripts/docker-entrypoint.sh
@@ -6,10 +6,10 @@ then
 fi
 
 if [[ ! -e ${PWD}/state.yml ]]; then
-  if [[ "${@}" == "blank_run_incremental" ]]; then
+  if [[ "${@}" == "lsst_run_incremental" ]]; then
     yesterday=$(date -d yesterday "+%d-%b-%Y %H:%M")
     echo "bookmarks:
-    blank_timestamp:
+    lsst_timestamp:
       last_record: $yesterday
 " > ${PWD}/state.yml
   else
