+++ scripts/lsst_run.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-COLLECTION="blank"
+COLLECTION="lsst"
 IMAGE="opencadc/${COLLECTION}2caom2"
 
 echo "Get a proxy certificate"
@@ -10,7 +10,7 @@ echo "Get image ${IMAGE}"
 docker pull ${IMAGE}
 
 echo "Run image ${IMAGE}"
-docker run --rm --name ${COLLECTION}_incremental  --user $(id -u):$(id -g) -e HOME=/usr/src/app -v ${PWD}:/usr/src/app/ ${IMAGE} ${COLLECTION}_run_incremental || exit $?
+docker run --rm --name ${COLLECTION}_todo  --user $(id -u):$(id -g) -e HOME=/usr/src/app -v ${PWD}:/usr/src/app/ ${IMAGE} ${COLLECTION}_run || exit $?
 
 date
 exit 0
